Move all git repo fixtures to
