+++ harness/tests/conftest.py
@@ -2,8 +2,10 @@
 
 from __future__ import annotations
 
+import shutil
 import subprocess
 import sys
+from collections.abc import Iterator
 from pathlib import Path
 from subprocess import PIPE
 from typing import Self
@@ -79,3 +81,69 @@ def git_repo(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> Path:
     monkeypatch.setattr(cli, "REPO_ROOT", tmp_path)
     monkeypatch.setattr(gate, "REPO_ROOT", tmp_path)
     return tmp_path
+
+
+@pytest.fixture
+def real_hook_repo(request: pytest.FixtureRequest, git_repo: Path, monkeypatch: pytest.MonkeyPatch) -> Path:
+    """Wire selected tracked hooks to a recorded executable in the disposable repository."""
+    hooks = git_repo / ".active-hooks"
+    hooks.mkdir()
+    for name in (*request.param, "_resolve"):
+        shutil.copy2(REPO_ROOT / ".githooks" / name, hooks / name)
+    gate.run_git(["config", "core.hooksPath", ".active-hooks"], git_repo)
+
+    executable = git_repo / "recorded-harness"
+    executable.write_text(
+        f"""#!{Path(sys.executable).as_posix()}
+import json
+import os
+import sys
+from pathlib import Path
+
+repo = Path.cwd()
+arguments = sys.argv[1:]
+command = arguments[0] if arguments else ""
+recorded = arguments.copy()
+if command == "prepare-commit-msg" and len(recorded) > 1:
+    recorded[1] = Path(recorded[1]).name
+with (repo / "harness.calls").open("a", encoding="utf-8") as handle:
+    handle.write(json.dumps({{"arguments": recorded, "RALPH_LOOP": os.environ.get("RALPH_LOOP")}}) + "\\n")
+real_file = repo / "harness.real"
+real_commands = real_file.read_text(encoding="utf-8").splitlines() if real_file.exists() else []
+if command == "prepare-commit-msg" or command in real_commands:
+    os.chdir({str(REPO_ROOT)!r})
+    from harness import cli, gate
+    os.chdir(repo)
+    gate.REPO_ROOT = repo
+    if command == "preflight":
+        gate.COMMIT_CHECKS = {{}}
+    cli.main(arguments)
+status_file = repo / "harness.exit"
+raise SystemExit(int(status_file.read_text(encoding="utf-8")) if status_file.exists() else 0)
+""",
+        encoding="utf-8",
+    )
+    executable.chmod(0o755)
+    (git_repo / ".git" / "harness-path").write_text(f"{executable}\n", encoding="utf-8")
+    monkeypatch.setenv("RALPH_LOOP", "1")
+    return git_repo
+
+
+def seed_repo(directory: Path) -> Path:
+    """Create a temp git repository with one commit and point gate's git calls at it."""
+    gate.run_git(["init", "-q"], directory)
+    gate.run_git(["config", "user.email", "harness@test.local"], directory)
+    gate.run_git(["config", "user.name", "harness-test"], directory)
+    (directory / "README.md").write_text("seed\n", encoding="utf-8")
+    gate.run_git(["add", "README.md"], directory)
+    gate.run_git(["commit", "-q", "-m", "seed"], directory)
+    return directory
+
+
+@pytest.fixture(scope="module")
+def scan_repo(tmp_path_factory: pytest.TempPathFactory) -> Iterator[Path]:
+    """A temp repo shared by the generated examples, since @given cannot take a per-test fixture."""
+    repo = seed_repo(tmp_path_factory.mktemp("banned-patterns"))
+    with pytest.MonkeyPatch.context() as patch:
+        patch.setattr(gate, "REPO_ROOT", repo)
+        yield repo
